+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Release 2.0.1 - December X, 2020
+### Fixed
+* BubbleChat: pass `avatar` prop value to prop `name` of Avatar, instead of `src`
+* Composer: remove defaultProps `tooltipAttachButton` and `tooltipSendButton`
+
 ## Release 2.0.0 - December 6, 2020
 ### Added
 * Migrate all from Got It Design System 1.12.11